+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "neo4j-vscode" extension will be documented in this file.
 
+## [0.1.5]
+
+- The existing result window will be used when running the same query
+- Result window is shown and results & errors are shown asyncronously
+- Added `CALL` and `YIELD` to Cypher keywords
+
 ## [0.1.4]
 
 - Added additional Cypher snippets for DMBS admin commands